requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cpu

accelerate==1.3.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anthropic==0.45.2
    # via -r requirements.in
anyio==4.8.0
    # via
    #   anthropic
    #   httpx
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via gtts
distro==1.9.0
    # via anthropic
dnspython==2.7.0
    # via pymongo
ffmpeg-python==0.2.0
    # via -r requirements.in
filelock==3.17.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.12.0
    # via
    #   huggingface-hub
    #   torch
future==1.0.0
    # via ffmpeg-python
gtts==2.5.4
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.26.0
    # via
    #   anthropic
    #   python-telegram-bot
huggingface-hub==0.28.0
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.5
    # via torch
jiter==0.8.2
    # via anthropic
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.2
    # via
    #   -r requirements.in
    #   accelerate
    #   transformers
packaging==24.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
psutil==6.1.1
    # via accelerate
pydantic==2.10.6
    # via anthropic
pydantic-core==2.27.2
    # via pydantic
pydub==0.25.1
    # via -r requirements.in
pymongo==4.6.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-telegram-bot==20.8
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   gtts
    #   huggingface-hub
    #   transformers
safetensors==0.5.2
    # via
    #   accelerate
    #   transformers
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
sympy==1.13.3
    # via torch
tokenizers==0.15.2
    # via transformers
torch==2.2.2
    # via
    #   -r requirements.in
    #   accelerate
    #   torchaudio
torchaudio==2.2.2
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.39.3
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via requests
